uorb: merge nuttx & posix files
