+++ src/modules/uORB/CMakeLists.txt
@@ -45,25 +45,8 @@ set(SRCS
 	Publication.cpp
 	Subscription.cpp
 	uORBManager.cpp
+	uORBDevices.cpp
 	)
-
-if(${OS} STREQUAL "nuttx")
-	list(APPEND SRCS
-		uORBDevices_nuttx.cpp
-		)
-elseif(${OS} STREQUAL "posix")
-	list(APPEND SRCS
-		uORBDevices_posix.cpp
-		)
-elseif(${OS} STREQUAL "posix-arm")
-	list(APPEND SRCS
-		uORBDevices_posix.cpp
-		)
-elseif(${OS} STREQUAL "qurt")
-	list(APPEND SRCS
-		uORBDevices_posix.cpp
-		)
-endif()
 
 px4_add_module(
 	MODULE modules__uORB
